Estimation of a SBM Laplace

This is the description of my function. Estimation of Stochastic Bloc Model under a Laplace emission law

VEM_SBM_laplace(Omega, Q, cl.init = kmeans(Omega, centers = Q, nstart =
  10)$cl, eps = 1e-05, maxIter = 50)

Arguments

Omega
a matrix: the data

Q
an integer: the number of clusters

cl.init
a initial clustering

eps
double convergence threshold

maxIter
integer the maximal number of itarates in the VEM algorithm

Examples

Lambda <- matrix(c(1, 2, 3, 2, 1, 4, 3, 4, 1), 3, 3)
p <- 200
alpha <- c(1/2, 1/4, 1/4)
mySBM <- rSBMLaplace(p, Lambda, alpha)
cl <- igraph::V(mySBM)$membership
Omega <- mySBM %>% igraph::as_adj(attr = "weight")
vBlocks <- 1:5
out <- lapply(vBlocks, function(q) VEM_SBM_laplace(Omega, q))